\name{BiocAnno2016-package}
\alias{BiocAnno2016-package}
\alias{BiocAnno2016}
\docType{package}
\title{
\packageTitle{BiocAnno2016}
}
\description{
\packageDescription{BiocAnno2016}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BiocAnno2016}
\packageIndices{BiocAnno2016}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BiocAnno2016}

Maintainer: \packageMaintainer{BiocAnno2016}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
