\name{getEset}
\alias{getEset}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to load an ExpressionSet for the annotation workshop
}
\description{
This function is called without any arguments to load data for the workshop.
}
\usage{
getEset()
}
%- maybe also 'usage' for other objects documented here.
}
\value{
No value is returned, but an ExpressionSet called 'eset' is loaded into
the user's workspace
}
\author{
James W. MacDonald
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip }% use one of  RShowDoc("KEYWORDS")

