\name{BiocAnno2017-package}
\alias{BiocAnno2017-package}
\alias{BiocAnno2017}
\docType{package}
\title{
\packageTitle{BiocAnno2017}
}
\description{
\packageDescription{BiocAnno2017}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BiocAnno2017}
\packageIndices{BiocAnno2017}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BiocAnno2017}

Maintainer: \packageMaintainer{BiocAnno2017}
}

\keyword{ package }
